+++ lib/rendering/effects/depth-fog-effect.ts
@@ -83,6 +83,8 @@ export class DepthFogEffect implements PostProcessEffect {
 		gl.bindVertexArray(this.emptyVao);
 		gl.drawArrays(gl.TRIANGLES, 0, 3);
 		gl.bindVertexArray(null);
+
+		ctx.stats.drawCalls++;
 	}
 
 	/**